refactor(cli): return the goblin mini-sprite to the transient left band (#336)
Restores the first iteration of #888 (adc65c0): the 13x3 goblin mini-sprite
lives in a transient extraRows band, left-aligned in the loop-stage gutter,
that claims rows only while the agent is mid-tool and releases them at idle —
popping up when work starts and clearing when it ends. Iterations 2-4 took it
through a one-row rail and then a constant right-aligned band (mascot-live +
mascot-band); after living with all three, the transient left band reads best:
a resting REPL pays no rows and no timer ticks, and the sprite's appearance
itself signals activity.

Mechanics: reverse-applied the delta adc65c0..557d6fc onto current main. The
only main-side overlap was the umber theme (kept). MascotBar returns with its
reserve-before-paint / erase-before-release ordering, alert dwell, and
co-tenancy with the bg bar and verdict rail; mascot-live.ts and mascot-band.ts
are removed. Gating is unchanged: AFK_GOBLIN_MASCOT=1, opt-in, default off.

Verification: lint, build, scan:env:check, audit:env:check, mascot suites
(58 tests), full interactive + _lib suites (1424 tests), test:pty (10).
