+++ src/cli/commands/interactive/repl-loop.ts
@@ -102,7 +102,7 @@ export async function runReplLoop(
       transcript,
       sigintHandler,
       suggestGhostEnabled,
-      { getLoopStageBar: () => footer?.loopStageBar, getLiveMascot: () => footer?.liveMascot },
+      { getLoopStageBar: () => footer?.loopStageBar, getMascotBar: () => footer?.mascotBar },
     );
 
     footer = setupFooterSubsystems(ctx, turnState);
@@ -133,19 +133,16 @@ export async function runReplLoop(
     footer?.bgResultNotifier.dispose();
     // Stop the footer painters top → bottom so each clears the exact row it
     // painted before the counts below it change. LoopStageBar positions from
-    // the full extraRows, so it must clear before mascotBand/bgStatusBar/
+    // the full extraRows, so it must clear before mascotBar/bgStatusBar/
     // verdictLedger shrink their counts. The mascot band positions from
     // bgBarRowCount + ledgerRowCount, so it must clear before those two. The
     // bg bar's clear row depends on the verdict count (its getAdjacentRows), so
     // it must clear before the verdict ledger drops ledgerRowCount to 0. The
     // verdict rail sits at the bottom (row N-1), independent of the others.
-    //
-    // The live mascot goes FIRST, ahead of that order: it owns no rows, but its
-    // animation ticker's only job is to ask the mascot band to repaint, so it
-    // must fall silent before that band releases its rows rather than after.
-    footer?.liveMascot.stop();
     footer?.loopStageBar.stop();
-    footer?.mascotBand.stop();
+    // The mascot band sits between the loop-stage rail and the bg bar, so it
+    // clears after the rail above it and before the counts below it change.
+    footer?.mascotBar.stop();
     footer?.bgStatusBar.stop();
     footer?.verdictLedger.stop();
     footer?.contextPane.dispose();
