+++ src/cli/commands/interactive/surface-setup.ts
@@ -14,7 +14,7 @@ import { onQuotaUpdate } from '../../../agent/quota-cache.js';
 import { formatStatusFields } from './shared.js';
 import type { TranscriptHandle } from './transcript.js';
 import type { LoopStageBar } from './loop-stage.js';
-import type { LiveMascot } from './mascot-live.js';
+import type { MascotBar } from './mascot-bar.js';
 import { buildPrompt, type TurnState } from './repl-loop-shared.js';
 
 /**
@@ -28,7 +28,7 @@ import { buildPrompt, type TurnState } from './repl-loop-shared.js';
 export interface SurfaceSetupDeps {
   getLoopStageBar: () => LoopStageBar | undefined;
   /** Same lazy-getter contract as {@link getLoopStageBar} (issue #336). */
-  getLiveMascot: () => LiveMascot | undefined;
+  getMascotBar: () => MascotBar | undefined;
 }
 
 export interface SurfaceSetupResult {
@@ -352,7 +352,7 @@ export async function setupSurface(
   // the bar exists. Matches the original hoisted-`let loopStageBar` capture.
   ctx.slashCtx.onStageChange = (stage, signals) => {
     deps.getLoopStageBar()?.repaint(stage);
-    deps.getLiveMascot()?.onStage(stage, signals);
+    deps.getMascotBar()?.onStage(stage, signals);
   };
   ctx.slashCtx.onContextProgress = async () => {
     await ctx.contextSampler.refresh();
